HUGE Celine code fusion, hope there is no regression
