+++ pod/video/forms.py
@@ -1410,13 +1410,21 @@ class ArchiveChoiceForm(forms.Form):
     if PROLONGATION_GRANTED:
         if DELETION_GRANTED:
             CHOICES = [
-                ("Extend", _("Extend (Automatically by %(rrd)s days)") % {"rrd": RALLONGE_RESPIT_DAYS}),
+                (
+                    "Extend",
+                    _("Extend (Automatically by %(rrd)s days)")
+                    % {"rrd": RALLONGE_RESPIT_DAYS},
+                ),
                 ("Archive", _("Archive")),
                 ("Delete", _("Delete")),
             ]
         else:
             CHOICES = [
-                ("Extend", _("Extend (Automatically by %(rrd)s days)") % {"rrd": RALLONGE_RESPIT_DAYS}),
+                (
+                    "Extend",
+                    _("Extend (Automatically by %(rrd)s days)")
+                    % {"rrd": RALLONGE_RESPIT_DAYS},
+                ),
                 ("Archive", _("Archive")),
             ]
     else:
@@ -1431,9 +1439,16 @@ class ArchiveChoiceForm(forms.Form):
             ]
 
     action_help_texts = {
-        "Extend": _("Select this option if you want to extend the publication of your video by %(rrd)s days.") % {"rrd": RALLONGE_RESPIT_DAYS},
-        "Archive": _("Select this option if you want to archive your video. It will then be unpublished and will no longer be accessible to the public. However, it will remain on the video server, along with its metadata, and its publication can be reactivated if necessary."),
-        "Delete": _("Select this option if you decide to permanently delete your video from the server, along with its metadata. You should only do this after careful consideration. It is therefore essential that you download the video and its metadata first."),
+        "Extend": _(
+            "Select this option if you want to extend the publication of your video by %(rrd)s days."
+        )
+        % {"rrd": RALLONGE_RESPIT_DAYS},
+        "Archive": _(
+            "Select this option if you want to archive your video. It will then be unpublished and will no longer be accessible to the public. However, it will remain on the video server, along with its metadata, and its publication can be reactivated if necessary."
+        ),
+        "Delete": _(
+            "Select this option if you decide to permanently delete your video from the server, along with its metadata. You should only do this after careful consideration. It is therefore essential that you download the video and its metadata first."
+        ),
     }
 
     action = forms.ChoiceField(
