+++ pod/video/management/commands/create_archive_package.py
@@ -3,27 +3,19 @@
 *  run with 'python manage.py create_archive_package [--dry]'
 """
 
-import csv
 import os
-import shutil
 
 from datetime import datetime, timedelta
-from defusedxml import minidom
 from django.conf import settings
 from django.core.mail import mail_managers
 from django.core.management.base import BaseCommand
-from django.core.serializers import serialize
 from django.template.defaultfilters import slugify, striptags
-from django.template.loader import render_to_string
 from django.utils.translation import activate
 from django.utils.translation import gettext as _
 
-from pod.settings import BASE_DIR
-from pod.video.models import Video, Notes, AdvancedNotes, Comment, ViewCount
-from pod.chapter.models import Chapter
-from pod.completion.models import Contributor, Document, Overlay, Track
-from pod.enrichment.models import Enrichment
+from pod.video.models import Video
 from pod.main.utils import sizeof_fmt
+from pod.video.utils import archive_pack, read_archived_csv
 
 """CUSTOM PARAMETERS."""
 LANGUAGE_CODE = getattr(settings, "LANGUAGE_CODE", "fr")
@@ -40,95 +32,6 @@ __TITLE_SITE__ = (
 )
 
 
-def store_as_dublincore(vid: Video, mediaPackage_dir: str, user_name: str) -> None:
-    """Store video metadata as Dublin Core Format in mediaPackage_dir."""
-    xmlcontent = '<?xml version="1.0" encoding="utf-8"?>\n'
-    xmlcontent += (
-        "<!DOCTYPE rdf:RDF PUBLIC " '"-//DUBLIN CORE//DCMES DTD 2002/07/31//EN" \n'
-    )
-    xmlcontent += (
-        '"http://dublincore.org/documents/2002/07' '/31/dcmes-xml/dcmes-xml-dtd.dtd">\n'
-    )
-    xmlcontent += (
-        "<rdf:RDF xmlns:rdf="
-        '"http://www.w3.org/1999/02/22-rdf-syntax-ns#"'
-        ' xmlns:dc ="http://purl.org/dc/elements/1.1/">\n'
-    )
-    rendered = render_to_string(
-        "videos/dublincore.html", {"video": vid, "xml": True}, None
-    )
-    xmlcontent += rendered
-    xmlcontent += "</rdf:RDF>"
-    # complete creator
-    mediaPackage_content = minidom.parseString(xmlcontent)
-
-    dc_creator = mediaPackage_content.getElementsByTagName("dc.creator")[0]
-
-    if dc_creator.firstChild is None:
-        new_node = mediaPackage_content.createTextNode(user_name)
-        dc_creator.appendChild(new_node)
-    else:
-        dc_creator.firstChild.replaceWholeText(user_name)
-
-    mediaPackage_file = os.path.join(mediaPackage_dir, "dublincore.xml")
-    with open(mediaPackage_file, "w") as f:
-        f.write(
-            minidom.parseString(
-                mediaPackage_content.toxml().replace("\n", "")
-            ).toprettyxml()
-        )
-
-
-def read_archived_csv() -> dict:
-    """Get data from ARCHIVE_CSV."""
-    csv_data = {}
-    if os.access(ARCHIVE_CSV, os.F_OK):
-        with open(ARCHIVE_CSV, "r", newline="", encoding="utf-8") as csvfile:
-            fieldnames = [
-                "Date",
-                "User name",
-                "User email",
-                "User Affiliation",
-                "User Establishment",
-                "Video Id",
-                "Video title",
-                "Video URL",
-                "Video type",
-                "Date added",
-            ]
-            reader = csv.DictReader(
-                csvfile, skipinitialspace=True, delimiter=";", fieldnames=fieldnames
-            )
-            for row in reader:
-                dico = {k: v for k, v in row.items()}
-                csv_data[dico["Video Id"]] = dico
-
-    return csv_data
-
-
-def archive_download_archive(slug):
-    """Generate a zip archive of the video and metadata from the concerned folder"""
-    path = os.path.join(os.path.dirname(BASE_DIR)) + "/pod/media/video_package"
-
-    cmd = Command()
-    cmd.archive_pack(slug, "", Video.objects.filter(slug=slug).first(), path, False)
-
-    zip_name = path + "/" + slug
-
-    mediaPackage_dir = os.path.join(path, "", slug)
-
-    directory_name = mediaPackage_dir
-
-    # Create 'path\to\zip_file.zip'
-    shutil.make_archive(zip_name, "zip", directory_name)
-
-    # remove old temp folder
-    path = os.path.join(mediaPackage_dir, "")
-    shutil.rmtree(path)
-
-    return "/media/video_package/" + slug + ".zip"
-
-
 class Command(BaseCommand):
     """Move old archived videos from disk to ARCHIVE_ROOT."""
 
@@ -143,117 +46,6 @@ class Command(BaseCommand):
             action="store_true",
             default=False,
         )
-
-    def export_complement(
-        self, folder: str, export_type: str, export_objects: list
-    ) -> None:
-        """Store a video complement as json."""
-        if len(export_objects) > 0:
-            export_file = os.path.join(folder, "%s.json" % export_type)
-            print("  * Export %s %s." % (len(export_objects), export_type))
-            if not self.dry_mode:
-                with open(export_file, "w") as out:
-                    content = serialize("json", export_objects)
-                    out.write(content)
-
-    def copy_archive_to(self, mediaPackage_dir: str, vid: Video) -> None:
-        """Move video source file to mediaPackage_dir."""
-        if os.access(vid.video.path, os.F_OK):
-            shutil.copy(
-                vid.video.path,
-                os.path.join(mediaPackage_dir, os.path.basename(vid.video.name)),
-            )
-        else:
-            print("ERROR: Cannot access to file '%s'." % vid.video.path)
-
-    def move_video_to_archive(self, mediaPackage_dir: str, vid: Video) -> None:
-        """Move video source file to mediaPackage_dir."""
-        if os.access(vid.video.path, os.F_OK):
-            print("  * Moving %s" % vid.video.path)
-            if not self.dry_mode:
-                shutil.move(
-                    vid.video.path,
-                    os.path.join(mediaPackage_dir, os.path.basename(vid.video.name)),
-                )
-                # Delete Video object
-                vid.delete()
-            # Deletes the video object and the associated folder (encoding, logs, etc.)
-            # Remove thumbnails (x3)
-        else:
-            print("ERROR: Cannot access to file '%s'." % vid.video.path)
-
-    def archive_pack(
-        self,
-        video_dir: str,
-        user_name: str,
-        vid: Video,
-        path_custom: str = "",
-        move_else_copy: bool = True,
-    ) -> None:
-        """Create a archive package for Video vid."""
-        # Get username from CSV
-        user_name = user_name.split("(")
-        # extract username from "Fullname (username)" string
-        user_name = user_name[-1][:-1]
-
-        # Create video folder
-        if path_custom == "":
-            mediaPackage_dir = os.path.join(ARCHIVE_ROOT, user_name, video_dir)
-        else:
-            mediaPackage_dir = os.path.join(path_custom, user_name, video_dir)
-
-        # Create directory to store all the data
-        os.makedirs(mediaPackage_dir, exist_ok=True)
-
-        # Move video file
-        store_as_dublincore(vid, mediaPackage_dir, user_name)
-
-        # Store Video complements as json
-        for model in [
-            Chapter,
-            Contributor,
-            Overlay,
-            Enrichment,
-            Notes,
-            AdvancedNotes,
-            Comment,
-            ViewCount,
-        ]:
-            # nb: contributors are already exported in dublincore.xml
-            self.export_complement(
-                mediaPackage_dir, model.__name__, model.objects.filter(video=vid)
-            )
-        # Export also the video itself as json
-        self.export_complement(mediaPackage_dir, "Video", [vid])
-
-        # Store also files linked to Enrichments
-        for enrich in Enrichment.objects.filter(video=vid):
-            if enrich.document:
-                print("  * Copying %s..." % enrich.document.file.path)
-                shutil.copy(enrich.document.file.path, mediaPackage_dir)
-            if enrich.image:
-                print("  * Copying %s..." % enrich.image.file.path)
-                shutil.copy(enrich.image.file.path, mediaPackage_dir)
-
-        # Store file complements.
-        for file in Document.objects.filter(video=vid):
-            print("  * Copying %s..." % file.document.file.path)
-            shutil.copy(file.document.file.path, mediaPackage_dir)
-
-        # Store additional tracks (caption / subtitles)
-        for track in Track.objects.filter(video=vid):
-            print("  * Copying %s..." % track.src.file.path)
-            shutil.copy(track.src.file.path, mediaPackage_dir)
-
-        # TODO:
-        # - Que faire du fichier CSV ? il faudrait y retirer toutes les
-        # lignes supprimées, quitte à faire un nouveau CSV
-
-        # You can decide if you simply copy the video or if you move it to the archive.
-        if move_else_copy:
-            self.move_video_to_archive(mediaPackage_dir, vid)
-        else:
-            self.copy_archive_to(mediaPackage_dir, vid)
 
     def get_list_video_html(self, list_video: list) -> str:
         """Generate an html version of list_video."""
@@ -301,17 +93,30 @@ class Command(BaseCommand):
                 continue
 
             # Recover original video slug
-            to_remove = len(_("Archived") + " 0000-00-00 ")
-            video_dir = "%04d-%s" % (vid.id, slugify(vid.title[to_remove:]))
-
             csv_entry = csv_data.get(str(vid.id))
+            to_remove = len(_("Archived") + " 0000-00-00 ")
+            video_dir_name = "%04d-%s" % (vid.id, slugify(vid.title[to_remove:]))
+            # Get username from CSV
+            user_name = csv_entry["User name"]
+            user_name = user_name.split("(")
+            # extract username from "Fullname (username)" string
+            user_name = user_name[-1][:-1]
+
+            media_package_dir = os.path.join(ARCHIVE_ROOT, user_name, video_dir_name)
+
             if csv_entry:
                 total_duration += vid.duration
                 total_processed += 1
                 if os.access(vid.video.path, os.F_OK):
                     total_weight += os.path.getsize(vid.video.path)
                 list_video.append(str(vid))
-                self.archive_pack(video_dir, csv_entry["User name"], vid)
+                archive_pack(
+                    media_package_dir,
+                    user_name,
+                    vid,
+                    only_copy=False,
+                    dry_mode=self.dry_mode,
+                )
             else:
                 print("  * Video %s not present in archived file" % vid.id)
             print("---")
