+++ pod/video/utils.py
@@ -7,16 +7,22 @@ import re
 import shutil
 import logging
 from math import ceil
+import csv
+from datetime import date
+from defusedxml import minidom
+from django.core.serializers import serialize
 
 from django.urls import reverse
 from django.conf import settings
 from django.http import JsonResponse
 from django.db.models import Q, Count
 from django.utils.translation import gettext_lazy as _
+from django.template.loader import render_to_string
+
 from pod.video_encode_transcript.models import EncodingVideo, EncodingAudio
 from pod.video_encode_transcript.models import PlaylistVideo
 from django.contrib.auth import get_user_model
-from .models import Video, Category, Type, Discipline
+from pod.video.models import Video, Category, Type, Discipline, VideoToDelete
 
 logger = logging.getLogger(__name__)
 User = get_user_model()
@@ -57,8 +63,15 @@ MANAGERS = getattr(settings, "MANAGERS", {})
 SECURE_SSL_REDIRECT = getattr(settings, "SECURE_SSL_REDIRECT", False)
 
 VIDEOS_DIR = getattr(settings, "VIDEOS_DIR", "videos")
+BASE_DIR = getattr(settings, "BASE_DIR", "/")
 
 NUMBER_TAGS_CLOUD = getattr(settings, "NUMBER_TAGS_CLOUD", 20)
+
+ARCHIVE_CSV = "%s/archived.csv" % settings.LOG_DIRECTORY
+
+ARCHIVE_OWNER_USERNAME = getattr(settings, "ARCHIVE_OWNER_USERNAME", "archive")
+
+ARCHIVE_ROOT = getattr(settings, "ARCHIVE_ROOT", "/video_archiving")
 
 ###############################################################
 # EMAIL
@@ -472,3 +485,297 @@ def get_filtered_owners_for_videos(user_videos, search_term=None, limit=20):
             :limit
         ]
     )
+
+
+def archive_video(vid):
+    """
+    It Allows the archive process without launching 'get_video archived deleted treatment' in the purpose to be used in other functions
+    """
+    write_in_csv(vid, "archived")
+    archive_user, created = User.objects.get_or_create(
+        username=ARCHIVE_OWNER_USERNAME,
+    )
+    # Rename video and change owner.
+    vid.owner = archive_user
+    vid.is_draft = True
+    vid.title = "%s %s %s" % (
+        _("Archived"),
+        date.today(),
+        vid.title,
+    )
+    # Trunc title to 250 chars max.
+    vid.title = vid.title[:250]
+    vid.save()
+
+    # add video to delete
+    vid_delete, created = VideoToDelete.objects.get_or_create(
+        date_deletion=vid.date_delete
+    )
+    vid_delete.video.add(vid)
+    vid_delete.save()
+
+
+def check_csv_header(csv_file: str, fieldnames: list) -> None:
+    """Check for (and add) missing columns in an existing CSV file."""
+    with open(csv_file, "r") as f:
+        lines = f.readlines()
+    if len(lines[0].split(";")) < len(fieldnames):
+        print("Adding missing header columns in %s." % csv_file)
+        lines[0] = ";".join(fieldnames) + "\n"
+        with open(csv_file, "w") as f:
+            f.writelines(lines)
+
+
+def write_in_csv(vid: Video, arch_type: str) -> None:
+    """Add in `type`.csv file informations about the video."""
+    file = "%s/%s.csv" % (settings.LOG_DIRECTORY, arch_type)
+    exists = os.path.isfile(file)
+
+    fieldnames = [
+        "Date",
+        "User name",
+        "User email",
+        "User Affiliation",
+        "User Establishment",
+        "Video Id",
+        "Video title",
+        "Video URL",
+        "Video type",
+        "Date added",
+        "Source file",
+        "Description",
+        "Views",
+    ]
+    if exists:
+        check_csv_header(file, fieldnames)
+
+    with open(file, "a", newline="", encoding="utf-8") as csvfile:
+        writer = csv.DictWriter(csvfile, delimiter=";", fieldnames=fieldnames)
+
+        if not exists:
+            writer.writeheader()
+
+        # Force the username attribute even if HIDE_USERNAME is true whereas the __str__ method
+        # of Owner Class used by vid.owner.owner doesn't do so
+        user_name = "%s %s (%s)" % (
+            vid.owner.first_name,
+            vid.owner.last_name,
+            vid.owner.username,
+        )
+
+        writer.writerow(
+            {
+                "Date": date.today(),
+                "User name": user_name,
+                "User email": vid.owner.email,
+                "User Affiliation": vid.owner.owner.affiliation,
+                "User Establishment": vid.owner.owner.establishment,
+                "Video Id": vid.id,
+                "Video title": vid.title,
+                "Video URL": "https:%s" % vid.get_full_url(),
+                "Video type": vid.type.title,
+                "Date added": "%s" % vid.date_added.strftime("%Y/%m/%d"),
+                "Source file": vid.video,
+                "Description": vid.description.replace(";", "$semic$")
+                .replace("\r", "")
+                .replace("\n\n", "\n")
+                .replace("\n", "$newl$"),
+                "Views": vid.viewcount,
+            }
+        )
+
+
+def store_as_dublincore(vid: Video, mediaPackage_dir: str, user_name: str) -> None:
+    """Store video metadata as Dublin Core Format in mediaPackage_dir."""
+    xmlcontent = '<?xml version="1.0" encoding="utf-8"?>\n'
+    xmlcontent += (
+        "<!DOCTYPE rdf:RDF PUBLIC " '"-//DUBLIN CORE//DCMES DTD 2002/07/31//EN" \n'
+    )
+    xmlcontent += (
+        '"http://dublincore.org/documents/2002/07' '/31/dcmes-xml/dcmes-xml-dtd.dtd">\n'
+    )
+    xmlcontent += (
+        "<rdf:RDF xmlns:rdf="
+        '"http://www.w3.org/1999/02/22-rdf-syntax-ns#"'
+        ' xmlns:dc ="http://purl.org/dc/elements/1.1/">\n'
+    )
+    rendered = render_to_string(
+        "videos/dublincore.html", {"video": vid, "xml": True}, None
+    )
+    xmlcontent += rendered
+    xmlcontent += "</rdf:RDF>"
+    # complete creator
+    mediaPackage_content = minidom.parseString(xmlcontent)
+
+    dc_creator = mediaPackage_content.getElementsByTagName("dc.creator")[0]
+
+    if dc_creator.firstChild is None:
+        new_node = mediaPackage_content.createTextNode(user_name)
+        dc_creator.appendChild(new_node)
+    else:
+        dc_creator.firstChild.replaceWholeText(user_name)
+
+    mediaPackage_file = os.path.join(mediaPackage_dir, "dublincore.xml")
+    with open(mediaPackage_file, "w") as f:
+        f.write(
+            minidom.parseString(
+                mediaPackage_content.toxml().replace("\n", "")
+            ).toprettyxml()
+        )
+
+
+def read_archived_csv() -> dict:
+    """Get data from ARCHIVE_CSV."""
+    csv_data = {}
+    if os.access(ARCHIVE_CSV, os.F_OK):
+        with open(ARCHIVE_CSV, "r", newline="", encoding="utf-8") as csvfile:
+            fieldnames = [
+                "Date",
+                "User name",
+                "User email",
+                "User Affiliation",
+                "User Establishment",
+                "Video Id",
+                "Video title",
+                "Video URL",
+                "Video type",
+                "Date added",
+            ]
+            reader = csv.DictReader(
+                csvfile, skipinitialspace=True, delimiter=";", fieldnames=fieldnames
+            )
+            for row in reader:
+                dico = {k: v for k, v in row.items()}
+                csv_data[dico["Video Id"]] = dico
+
+    return csv_data
+
+
+def export_complement(
+    folder: str, export_type: str, export_objects: list, dry_mode: bool = True
+) -> None:
+    """Store a video complement as json."""
+    if len(export_objects) > 0:
+        export_file = os.path.join(folder, "%s.json" % export_type)
+        print("  * Export %s %s." % (len(export_objects), export_type))
+        if not dry_mode:
+            with open(export_file, "w") as out:
+                content = serialize("json", export_objects)
+                out.write(content)
+
+
+def move_video_to_archive(
+    mediaPackage_dir: str, vid: Video, dry_mode: bool = True
+) -> None:
+    """Move video source file to mediaPackage_dir."""
+    if os.access(vid.video.path, os.F_OK):
+        print(
+            "  * Moving %s to " % vid.video.path,
+            os.path.join(mediaPackage_dir, os.path.basename(vid.video.name)),
+        )
+        if not dry_mode:
+            shutil.move(
+                vid.video.path,
+                os.path.join(mediaPackage_dir, os.path.basename(vid.video.name)),
+            )
+            # Delete Video object
+            vid.delete()
+        # Deletes the video object and the associated folder (encoding, logs, etc.)
+        # Remove thumbnails (x3)
+    else:
+        print("ERROR: Cannot access to file '%s'." % vid.video.path)
+
+
+def copy_archive_to(media_package_dir: str, vid: Video) -> None:
+    """Move video source file to mediaPackage_dir."""
+    if os.access(vid.video.path, os.F_OK):
+        shutil.copy(
+            vid.video.path,
+            os.path.join(media_package_dir),
+        )
+    else:
+        print("ERROR: Cannot access to file '%s'." % vid.video.path)
+
+
+def archive_pack(
+    media_package_dir: str,
+    user_name: str,
+    vid: Video,
+    only_copy: bool = True,
+    dry_mode: bool = True,
+) -> None:
+    """Create a archive package for Video vid."""
+    from pod.video.models import Notes, AdvancedNotes, Comment, ViewCount
+    from pod.chapter.models import Chapter
+    from pod.completion.models import Contributor, Document, Overlay, Track
+    from pod.enrichment.models import Enrichment
+
+    # Create directory to store all the data
+    os.makedirs(media_package_dir, exist_ok=True)
+
+    # Move video file
+    store_as_dublincore(vid, media_package_dir, user_name)
+
+    # Store Video complements as json
+    for model in [
+        Chapter,
+        Contributor,
+        Overlay,
+        Enrichment,
+        Notes,
+        AdvancedNotes,
+        Comment,
+        ViewCount,
+    ]:
+        # nb: contributors are already exported in dublincore.xml
+        export_complement(
+            media_package_dir, model.__name__, model.objects.filter(video=vid), dry_mode
+        )
+    # Export also the video itself as json
+    export_complement(media_package_dir, "Video", [vid], dry_mode)
+
+    # Store also files linked to Enrichments
+    for enrich in Enrichment.objects.filter(video=vid):
+        if enrich.document:
+            print("  * Copying %s..." % enrich.document.file.path)
+            shutil.copy(enrich.document.file.path, media_package_dir)
+        if enrich.image:
+            print("  * Copying %s..." % enrich.image.file.path)
+            shutil.copy(enrich.image.file.path, media_package_dir)
+
+    # Store file complements.
+    for file in Document.objects.filter(video=vid):
+        print("  * Copying %s..." % file.document.file.path)
+        shutil.copy(file.document.file.path, media_package_dir)
+
+    # Store additional tracks (caption / subtitles)
+    for track in Track.objects.filter(video=vid):
+        print("  * Copying %s..." % track.src.file.path)
+        shutil.copy(track.src.file.path, media_package_dir)
+
+    # TODO:
+    # - Que faire du fichier CSV ? il faudrait y retirer toutes les
+    # lignes supprimées, quitte à faire un nouveau CSV
+
+    # You can decide if you simply copy the video or if you move it to the archive.
+    if only_copy:
+        copy_archive_to(media_package_dir, vid)
+    else:
+        move_video_to_archive(media_package_dir, vid, dry_mode)
+
+
+def archive_and_get_link(slug):
+    """Generate a zip archive of the video and metadata from the concerned folder"""
+    media_url = getattr(settings, "MEDIA_URL", "/media/")
+    media_root = getattr(settings, "MEDIA_ROOT", os.path.join(BASE_DIR, "media"))
+
+    media_package_dir = os.path.join(media_root, "tmp", slug)
+    vid = Video.objects.filter(slug=slug).first()
+    archive_pack(media_package_dir, "", vid, only_copy=True, dry_mode=False)
+
+    shutil.make_archive(media_package_dir, "zip", media_package_dir)
+
+    # remove old temp folder
+    shutil.rmtree(media_package_dir)
+
+    return media_url + "/tmp/" + slug + ".zip"
